--- src/main/java/org/matsim/run/RunBerlinScenarioWithBicycles.java
@@ -1,158 +1,0 @@
-/* *********************************************************************** *
- * project: org.matsim.*
- *                                                                         *
- * *********************************************************************** *
- *                                                                         *
- * copyright       : (C) 2017 by the members listed in the COPYING,        *
- *                   LICENSE and WARRANTY file.                            *
- * email           : info at matsim dot org                                *
- *                                                                         *
- * *********************************************************************** *
- *                                                                         *
- *   This program is free software; you can redistribute it and/or modify  *
- *   it under the terms of the GNU General Public License as published by  *
- *   the Free Software Foundation; either version 2 of the License, or     *
- *   (at your option) any later version.                                   *
- *   See also COPYING, LICENSE and WARRANTY file                           *
- *                                                                         *
- * *********************************************************************** */
-
-package org.matsim.run;
-
-import ch.sbb.matsim.routing.pt.raptor.SwissRailRaptorModule;
-import org.apache.log4j.Logger;
-import org.matsim.api.core.v01.Scenario;
-import org.matsim.api.core.v01.TransportMode;
-import org.matsim.contrib.bicycle.BicycleConfigGroup;
-import org.matsim.contrib.bicycle.Bicycles;
-import org.matsim.core.config.Config;
-import org.matsim.core.config.ConfigGroup;
-import org.matsim.core.config.ConfigUtils;
-import org.matsim.core.config.groups.PlanCalcScoreConfigGroup.ActivityParams;
-import org.matsim.core.config.groups.QSimConfigGroup.TrafficDynamics;
-import org.matsim.core.config.groups.VspExperimentalConfigGroup;
-import org.matsim.core.controler.AbstractModule;
-import org.matsim.core.controler.Controler;
-import org.matsim.core.controler.OutputDirectoryLogging;
-import org.matsim.core.gbl.Gbl;
-import org.matsim.core.scenario.ScenarioUtils;
-
-import java.util.Arrays;
-
-import static org.matsim.core.config.groups.ControlerConfigGroup.RoutingAlgorithmType.FastAStarLandmarks;
-
-/**
-* @author ikaddoura, dziemke
-*/
-
-public final class RunBerlinScenarioWithBicycles {
-
-	private static final Logger log = Logger.getLogger(RunBerlinScenarioWithBicycles.class );
-
-	public static void main(String[] args) {
-		
-		for (String arg : args) {
-			log.info( arg );
-		}
-		
-		if ( args.length==0 ) {
-			throw new IllegalArgumentException("Config file needs to be provided.");
-		}
-
-		BicycleConfigGroup bicycleConfigGroup = new BicycleConfigGroup();
-		bicycleConfigGroup.setBicycleMode("bicycle");
-
-		Config config = prepareConfig( args, bicycleConfigGroup ) ;
-		Scenario scenario = prepareScenario( config ) ;
-		Controler controler = prepareControler( scenario ) ;
-		controler.run() ;
-	}
-
-	public static Controler prepareControler( Scenario scenario ) {
-		// note that for something like signals, and presumably drt, one needs the controler object
-		
-		Gbl.assertNotNull(scenario);
-		
-		final Controler controler = new Controler( scenario );
-		
-		if (controler.getConfig().transit().isUsingTransitInMobsim()) {
-			// use the sbb pt raptor router
-			controler.addOverridingModule( new AbstractModule() {
-				@Override
-				public void install() {
-					install( new SwissRailRaptorModule() );
-				}
-			} );
-		} else {
-			log.warn("Public transit will be teleported and not simulated in the mobsim! "
-					+ "This will have a significant effect on pt-related parameters (travel times, modal split, and so on). "
-					+ "Should only be used for testing or car-focused studies with a fixed modal split.  ");
-		}
-		
-		// use the (congested) car travel time for the teleported ride mode
-		controler.addOverridingModule( new AbstractModule() {
-			@Override
-			public void install() {
-				addTravelTimeBinding( TransportMode.ride ).to( networkTravelTime() );
-				addTravelDisutilityFactoryBinding( TransportMode.ride ).to( carTravelDisutilityFactoryKey() );
-			}
-		} );
-
-		Bicycles.addAsOverridingModule(controler);
-
-		return controler;
-	}
-	
-	public static Scenario prepareScenario( Config config ) {
-		Gbl.assertNotNull( config );
-		
-		// note that the path for this is different when run from GUI (path of original config) vs.
-		// when run from command line/IDE (java root).  :-(    See comment in method.  kai, jul'18
-		// yy Does this comment still apply?  kai, jul'19
-
-		final Scenario scenario = ScenarioUtils.loadScenario( config );
-
-		return scenario;
-	}
-	
-	public static Config prepareConfig( String [] args, ConfigGroup... customModules ) {
-		OutputDirectoryLogging.catchLogEntries();
-		
-		String[] typedArgs = Arrays.copyOfRange( args, 1, args.length );
-
-		final Config config = ConfigUtils.loadConfig( args[ 0 ], customModules ); // I need this to set the context
-		
-		config.controler().setRoutingAlgorithmType( FastAStarLandmarks );
-		
-		config.subtourModeChoice().setProbaForRandomSingleTripMode( 0.5 );
-		
-		config.plansCalcRoute().setRoutingRandomness( 3. );
-		config.plansCalcRoute().removeModeRoutingParams(TransportMode.ride);
-		config.plansCalcRoute().removeModeRoutingParams(TransportMode.pt);
-		config.plansCalcRoute().removeModeRoutingParams(TransportMode.bike);
-		config.plansCalcRoute().removeModeRoutingParams("bicycle");
-		config.plansCalcRoute().removeModeRoutingParams("undefined");
-	
-		config.qsim().setInsertingWaitingVehiclesBeforeDrivingVehicles( true );
-				
-		// vsp defaults
-		config.vspExperimental().setVspDefaultsCheckingLevel( VspExperimentalConfigGroup.VspDefaultsCheckingLevel.info );
-		config.plansCalcRoute().setInsertingAccessEgressWalk( true );
-		config.qsim().setUsingTravelTimeCheckInTeleportation( true );
-		config.qsim().setTrafficDynamics( TrafficDynamics.kinematicWaves );
-				
-		// activities:
-		for ( long ii = 600 ; ii <= 97200; ii+=600 ) {
-			config.planCalcScore().addActivityParams( new ActivityParams( "home_" + ii + ".0" ).setTypicalDuration( ii ) );
-			config.planCalcScore().addActivityParams( new ActivityParams( "work_" + ii + ".0" ).setTypicalDuration( ii ).setOpeningTime(6. * 3600. ).setClosingTime(20. * 3600. ) );
-			config.planCalcScore().addActivityParams( new ActivityParams( "leisure_" + ii + ".0" ).setTypicalDuration( ii ).setOpeningTime(9. * 3600. ).setClosingTime(27. * 3600. ) );
-			config.planCalcScore().addActivityParams( new ActivityParams( "shopping_" + ii + ".0" ).setTypicalDuration( ii ).setOpeningTime(8. * 3600. ).setClosingTime(20. * 3600. ) );
-			config.planCalcScore().addActivityParams( new ActivityParams( "other_" + ii + ".0" ).setTypicalDuration( ii ) );
-		}
-		config.planCalcScore().addActivityParams( new ActivityParams( "freight" ).setTypicalDuration( 12.*3600. ) );
-
-		ConfigUtils.applyCommandline( config, typedArgs ) ;
-
-		return config ;
-	}
-}
